refactor : Line의 Set<LineStation> 일급 컬렉션으로 변경
