+++ src/main/java/wooteco/subway/admin/domain/Line.java
@@ -3,13 +3,12 @@ package wooteco.subway.admin.domain;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.Set;
 
 import org.springframework.data.annotation.Id;
+import org.springframework.data.relational.core.mapping.Embedded;
 
 public class Line {
 	@Id
@@ -21,7 +20,8 @@ public class Line {
 	private String bgColor;
 	private LocalDateTime createdAt;
 	private LocalDateTime updatedAt;
-	private Set<LineStation> stations = new HashSet<>();
+	@Embedded.Empty
+	private LineStations lineStations;
 
 	public Line() {
 	}
@@ -35,10 +35,40 @@ public class Line {
 		this.bgColor = bgColor;
 		this.createdAt = LocalDateTime.now();
 		this.updatedAt = LocalDateTime.now();
+		this.lineStations = new LineStations();
 	}
 
 	public Line(String name, LocalTime startTime, LocalTime endTime, int intervalTime) {
 		this(null, name, startTime, endTime, intervalTime, "bg-red-400");
+	}
+
+	public void update(Line line) {
+		if (line.getName() != null) {
+			this.name = line.getName();
+		}
+		if (line.getStartTime() != null) {
+			this.startTime = line.getStartTime();
+		}
+		if (line.getEndTime() != null) {
+			this.endTime = line.getEndTime();
+		}
+		if (line.getIntervalTime() != 0) {
+			this.intervalTime = line.getIntervalTime();
+		}
+
+		this.updatedAt = LocalDateTime.now();
+	}
+
+	public void addLineStation(LineStation lineStation) {
+		lineStations.addLineStation(lineStation);
+	}
+
+	public void removeLineStationById(Long stationId) {
+		lineStations.removeLineStationById(stationId);
+	}
+
+	public List<Long> getLineStationsId() {
+		return lineStations.getLineStationsId();
 	}
 
 	public Long getId() {
@@ -61,8 +91,8 @@ public class Line {
 		return intervalTime;
 	}
 
-	public Set<LineStation> getStations() {
-		return stations;
+	public LineStations getLineStations() {
+		return lineStations;
 	}
 
 	public LocalDateTime getCreatedAt() {
@@ -75,74 +105,5 @@ public class Line {
 
 	public String getBgColor() {
 		return this.bgColor;
-	}
-
-	public void update(Line line) {
-		if (line.getName() != null) {
-			this.name = line.getName();
-		}
-		if (line.getStartTime() != null) {
-			this.startTime = line.getStartTime();
-		}
-		if (line.getEndTime() != null) {
-			this.endTime = line.getEndTime();
-		}
-		if (line.getIntervalTime() != 0) {
-			this.intervalTime = line.getIntervalTime();
-		}
-
-		this.updatedAt = LocalDateTime.now();
-	}
-
-	public void addLineStation(LineStation lineStation) {
-		stations.stream()
-			.filter(it -> Objects.equals(it.getPreStationId(), lineStation.getPreStationId()))
-			.findAny()
-			.ifPresent(it -> it.updatePreLineStation(lineStation.getStationId()));
-
-		stations.add(lineStation);
-	}
-
-	public void removeLineStationById(Long stationId) {
-		LineStation targetLineStation = stations.stream()
-			.filter(it -> Objects.equals(it.getStationId(), stationId))
-			.findFirst()
-			.orElseThrow(RuntimeException::new);
-
-		stations.stream()
-			.filter(it -> Objects.equals(it.getPreStationId(), stationId))
-			.findFirst()
-			.ifPresent(it -> it.updatePreLineStation(targetLineStation.getPreStationId()));
-
-		stations.remove(targetLineStation);
-	}
-
-	public List<Long> getLineStationsId() {
-		if (stations.isEmpty()) {
-			return new ArrayList<>();
-		}
-
-		LineStation firstLineStation = stations.stream()
-			.filter(it -> it.getPreStationId() == null)
-			.findFirst()
-			.orElseThrow(RuntimeException::new);
-
-		List<Long> stationIds = new ArrayList<>();
-		stationIds.add(firstLineStation.getStationId());
-
-		while (true) {
-			Long lastStationId = stationIds.get(stationIds.size() - 1);
-			Optional<LineStation> nextLineStation = stations.stream()
-				.filter(it -> Objects.equals(it.getPreStationId(), lastStationId))
-				.findFirst();
-
-			if (!nextLineStation.isPresent()) {
-				break;
-			}
-
-			stationIds.add(nextLineStation.get().getStationId());
-		}
-
-		return stationIds;
 	}
 }
