+++ src/main/java/wooteco/subway/admin/domain/LineStations.java
@@ -1,0 +1,73 @@
+package wooteco.subway.admin.domain;
+
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
+
+/**
+ *   class description
+ *
+ *   @author ParkDooWon
+ */
+public class LineStations {
+	private Set<LineStation> stations = new HashSet<>();
+
+	public void addLineStation(LineStation lineStation) {
+		stations.stream()
+			.filter(it -> Objects.equals(it.getPreStationId(), lineStation.getPreStationId()))
+			.findAny()
+			.ifPresent(it -> it.updatePreLineStation(lineStation.getStationId()));
+
+		stations.add(lineStation);
+	}
+
+	public void removeLineStationById(Long stationId) {
+		LineStation targetLineStation = stations.stream()
+			.filter(it -> Objects.equals(it.getStationId(), stationId))
+			.findFirst()
+			.orElseThrow(RuntimeException::new);
+
+		stations.stream()
+			.filter(it -> Objects.equals(it.getPreStationId(), stationId))
+			.findFirst()
+			.ifPresent(it -> it.updatePreLineStation(targetLineStation.getPreStationId()));
+
+		stations.remove(targetLineStation);
+	}
+
+	public List<Long> getLineStationsId() {
+		if (stations.isEmpty()) {
+			return new ArrayList<>();
+		}
+
+		LineStation firstLineStation = stations.stream()
+			.filter(it -> it.getPreStationId() == null)
+			.findFirst()
+			.orElseThrow(RuntimeException::new);
+
+		List<Long> stationIds = new ArrayList<>();
+		stationIds.add(firstLineStation.getStationId());
+
+		while (true) {
+			Long lastStationId = stationIds.get(stationIds.size() - 1);
+			Optional<LineStation> nextLineStation = stations.stream()
+				.filter(it -> Objects.equals(it.getPreStationId(), lastStationId))
+				.findFirst();
+
+			if (!nextLineStation.isPresent()) {
+				break;
+			}
+
+			stationIds.add(nextLineStation.get().getStationId());
+		}
+
+		return stationIds;
+	}
+
+	public Set<LineStation> getStations() {
+		return stations;
+	}
+}
