+++ src/main/resources/templates/service/index.html
@@ -32,10 +32,10 @@
           </li>
           <br />
           <li class="inline-block p-1">
-            <a href="map.html" class="underline">전체 노선 보기</a>
+            <a href="/map-page" class="underline">전체 노선 보기</a>
           </li>
           <li class="inline-block p-1">
-            <a href="search.html" class="underline">경로 검색</a>
+            <a href="/search-page" class="underline">경로 검색</a>
           </li>
         </ul>
       </div>
